add separate vendor logic into model
